refactor: comprehensive architecture improvements across all crates
- Centralize external dependencies in workspace Cargo.toml
- Add debug_assert guards before all unsafe String::from_utf8_unchecked calls
- Replace String error erasure with typed ScopesError in ParseError chain
- Add #[inline] to hot-path functions across codec, scopes, sourcemap, generator
- Add Hash, PartialOrd, Ord trait impls to Segment for completeness
- Add single-char fast path to vlq_decode_unsigned
- Use Box<RawValue> for indexed source map sections (avoid JSON re-serialize)
- Add get_source/get_name safe accessor methods to SourceMap
- Use impl Into<String> for flexible generator setter APIs
- Extract resolve_sources/build_source_map helpers to eliminate duplication
- Remove redundant dedup HashMaps from ConcatBuilder
- Cache has_range_mappings boolean to avoid O(n) scan on every call
- Convert find_nth_scope from recursive to iterative (stack overflow safety)
- Change scopes definition field from usize to u32 for platform consistency
- Replace .unwrap() with .expect() with descriptive messages
- Fix CLI error handling: replace unwrap with proper CliError propagation

+++ packages/scopes-wasm/src/lib.rs
@@ -142,7 +142,7 @@ fn encode_generated_range(range: &srcmap_scopes::GeneratedRange) -> JsValue {
     js_sys::Reflect::set(&obj, &"isHidden".into(), &range.is_hidden.into()).unwrap_or(false);
 
     let def_val: JsValue = match range.definition {
-        Some(d) => (d as u32).into(),
+        Some(d) => d.into(),
         None => JsValue::NULL,
     };
     js_sys::Reflect::set(&obj, &"definition".into(), &def_val).unwrap_or(false);
@@ -246,7 +246,7 @@ struct GeneratedRangeJson {
     #[serde(default, rename = "isHidden")]
     is_hidden: bool,
     #[serde(default)]
-    definition: Option<usize>,
+    definition: Option<u32>,
     #[serde(default, rename = "callSite")]
     call_site: Option<CallSiteJson>,
     #[serde(default)]
